fix: fixed memory access and initial refactor
