+++ src/messages/index.ts
@@ -1,3 +1,2 @@
 export { AuthMessageTypes } from "./auth";
-export { MailMessageTypes } from "./mail";
 export { Domain } from "./erc712";
