--- src/messages/mail.ts
@@ -1,14 +1,0 @@
-// The named list of all type definitions
-const MailMessageTypes = {
-  Person: [
-    { name: "name", type: "string" },
-    { name: "wallet", type: "address" },
-  ],
-  Mail: [
-    { name: "from", type: "Person" },
-    { name: "to", type: "Person" },
-    { name: "contents", type: "string" },
-  ],
-};
-
-export { MailMessageTypes };
