+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export { signAuthMessage } from "./signAuthMessage";
-export { signMailMessage } from "./signMailMessage";
 export { getSignerFromMnemonic, getSignerFromPrivateKey } from "./signer";
 export { packParameters } from "./packParameters";
