+++ tasks/deploy/auth.ts
@@ -1,14 +1,16 @@
 import { task } from "hardhat/config";
 import { TaskArguments } from "hardhat/types";
 
-import { AuthVerifier } from "../../src/types/AuthVerifier";
-import { AuthVerifier__factory } from "../../src/types/factories/AuthVerifier__factory";
+import { AccessTokenVerifier } from "../../src/types/AccessTokenVerifier";
+import { AccessTokenVerifier__factory } from "../../src/types/factories/AccessTokenVerifier__factory";
 
 task("deploy:Auth")
   .addParam("root", "Root key")
   .setAction(async function (taskArguments: TaskArguments, { ethers }) {
-    const authFactory: AuthVerifier__factory = <AuthVerifier__factory>await ethers.getContractFactory("AuthVerifier");
-    const auth: AuthVerifier = <AuthVerifier>await authFactory.deploy(taskArguments.root);
+    const authFactory: AccessTokenVerifier__factory = <AccessTokenVerifier__factory>(
+      await ethers.getContractFactory("AccessTokenVerifier")
+    );
+    const auth: AccessTokenVerifier = <AccessTokenVerifier>await authFactory.deploy(taskArguments.root);
     await auth.deployed();
     console.log("Auth deployed to: ", auth.address);
   });
