+++ test/KeyInfrastructure/KeyInfrastructure.behaviour.ts
@@ -1,32 +1,15 @@
 import { artifacts, ethers, waffle } from "hardhat";
 import type { Artifact } from "hardhat/types";
 import type { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
-import { Signers } from "./types";
-import { KeyInfrastructure } from "../src/types/KeyInfrastructure";
+import { Signers } from "../types";
+import { KeyInfrastructure } from "../../src/types/KeyInfrastructure";
 
 const chai = require("chai");
 const { solidity } = waffle;
 chai.use(solidity);
 const { expect } = chai;
 
-describe("Key Infrastructure", function () {
-  before(async function () {
-    this.signers = {} as Signers;
-
-    const signers: SignerWithAddress[] = await ethers.getSigners();
-    this.signers.admin = signers[0];
-    this.signers.user0 = signers[1];
-    this.signers.user1 = signers[2];
-    this.signers.user2 = signers[3];
-  });
-
-  before("deploy new", async function () {
-    const keyInfraArtifact: Artifact = await artifacts.readArtifact("KeyInfrastructure");
-    this.keyInfrastructure = <KeyInfrastructure>(
-      await waffle.deployContract(this.signers.admin, keyInfraArtifact, [this.signers.admin.address])
-    );
-  });
-
+const shouldBehaveLikeKeyInfrastructure = function () {
   it("key infrastructure should have been initialised correctly", async function () {
     expect(await this.keyInfrastructure.callStatic.getRootKey()).to.equal(this.signers.admin.address);
   });
@@ -71,4 +54,6 @@ describe("Key Infrastructure", function () {
       expect(await this.keyInfrastructure.callStatic.getIssuerKey()).to.equal(this.signers.user0.address);
     });
   });
-});
+};
+
+export { shouldBehaveLikeKeyInfrastructure };
